requirements.txt
torch==2.0.1
transformers==4.30.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
yfinance==0.2.18
alpha-vantage==2.3.1
scikit-learn==1.3.0
ta==0.10.2
schedule==1.2.0
python-dotenv==1.0.0
huggingface-hub==0.16.4
datasets==2.14.0
accelerate==0.20.3
